pyproject.toml
[project]
name = "cse144-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
